IGAPP-173: upgrade flow-types
