+++ web/flow-typed/npm/@babel/polyfill_v7.x.x.js
@@ -1,4 +1,4 @@
-// flow-typed signature: 70ae65c95b3420e3469c9c30171c1835
+// flow-typed signature: 2020db9e199604de6d1c9bf881778aab
 // flow-typed version: c6154227d1/@babel/polyfill_v7.x.x/flow_>=v0.104.x
 
 declare module '@babel/polyfill' {}
