+++ web/flow-typed/npm/flow-bin_v0.x.x.js
@@ -1,0 +1,6 @@
+// flow-typed signature: defcb6d3fbb0946084b6113e14859c31
+// flow-typed version: c6154227d1/flow-bin_v0.x.x/flow_>=v0.104.x
+
+declare module "flow-bin" {
+  declare module.exports: string;
+}
