+++ web/flow-typed/npm/moment_v2.x.x.js
@@ -1,5 +1,5 @@
-// flow-typed signature: 0d126f47f5358304e3a7f505a3ad94d2
-// flow-typed version: 8687adfbcb/moment_v2.x.x/flow_>=v0.104.x
+// flow-typed signature: 2c015b83e2d690a650b43f5190851b73
+// flow-typed version: e7ad2464da/moment_v2.x.x/flow_>=v0.104.x
 
 type moment$MomentOptions = {
   y?: number | string,
@@ -144,6 +144,12 @@ declare class moment$Moment {
   static (string: ?string, format: string | Array<string>): moment$Moment;
   static (
     string: ?string,
+    // Support for strict string parsing without format works since moment v2.25.0
+    // https://github.com/moment/moment/issues/2469
+    strict: boolean
+  ): moment$Moment;
+  static (
+    string: ?string,
     format: string | Array<string>,
     strict: boolean
   ): moment$Moment;
@@ -172,6 +178,12 @@ declare class moment$Moment {
       | void
   ): moment$Moment;
   static utc(string: string, format: string | Array<string>): moment$Moment;
+  static utc(
+    string: string,
+    // Support for strict string parsing without format works since moment v2.25.0
+    // https://github.com/moment/moment/issues/2469
+    strict: boolean
+  ): moment$Moment;
   static utc(
     string: string,
     format: string | Array<string>,
