+++ web/flow-typed/npm/redux_v4.x.x.js
@@ -1,4 +1,4 @@
-// flow-typed signature: f62df6dbce399d55b0f2954c5ac1bd4e
+// flow-typed signature: 52b77187072bb7987eb881f4e7ae40cb
 // flow-typed version: c6154227d1/redux_v4.x.x/flow_>=v0.104.x
 
 declare module 'redux' {
