+++ web/package.json
@@ -110,8 +110,8 @@
     "enzyme-to-json": "^3.4.3",
     "es-check": "^5.2.0",
     "file-loader": "^6.2.0",
-    "flow-bin": "0.140.0",
-    "flow-typed": "^2.5.1",
+    "flow-bin": "0.141.0",
+    "flow-typed": "^3.2.1",
     "html-loader": "^1.3.2",
     "html-webpack-plugin": "^4.5.1",
     "htmlparser2": "^6.0.0",
